More aggressive block spawning
Still careful to wait for "ready" so that the engine doesn't get
overwhelmed. There is the risk of deleting an object before it is
ready, but as long as max_on_screen is larger than spawn_per_tick it
should be fine. And that shouldn't really be possible given the
requency of message passing back and fourth between the page thread and
the worker.